+++ tests/TestCase/FilesystemTest.php
@@ -272,6 +272,23 @@ class FilesystemTest extends TestCase
     }
 
     /**
+     * Test for `makePathRelative()` method
+     * @test
+     */
+    public function testMakePathRelative(): void
+    {
+        $endPath = DS . 'a' . DS . 'b' . DS . 'c' . DS . 'd';
+        $startPath = DS . 'a' . DS . 'b';
+        $expected = 'c' . DS . 'd';
+
+        $this->assertSame($expected, Filesystem::instance()->makePathRelative($endPath, $startPath));
+        $this->assertSame($expected, Filesystem::instance()->makePathRelative($endPath . DS, $startPath . DS));
+        $this->assertSame($expected, Filesystem::instance()->makePathRelative($endPath, $startPath . DS));
+        $this->assertSame($expected, Filesystem::instance()->makePathRelative($endPath . DS, $startPath));
+        $this->assertSame($expected . DS . 'file.php', Filesystem::instance()->makePathRelative($startPath . DS . 'c' . DS . 'd' . DS . 'file.php', $startPath));
+    }
+
+    /**
      * Test for `normalizePath()` method
      * @test
      */