+++ CHANGELOG.md
@@ -4,9 +4,9 @@
 * `Exceptionist::__callStatic()` now handles calls containing with the "Not"
     word (e.g. `isNotArray()` or `fileNotExists()`);
 * added `EventAssertTrait::assertEventFiredWithArgs()` method;
+* added `Filesystem::makePathRelative()` method;
 * added `Exceptionist::isFalse()` method;
 * fixed little bug for `Filesystem::rtr()`;
-* improved `EventAssertTrait::assertEventFiredWithArgs()` method;
 * many methods already supported by `Exceptionist` have been documented.
 
 ### 1.5.9